+++ src/ui/components/util/barrel.ts
@@ -1,2 +1,2 @@
 export * from './PlayerContext';
-export * from './Localization';
+export * from './Locale';
